chart

Used to add missing data to input files and customize format

chart(
  data = NULL,
  col_agg = "class",
  chart_type = "all",
  aspect_ratio = 0.75,
  size = 0.75,
  size_text = 15,
  theme = NULL,
  ncol = NULL,
  scales = "free_y",
  save = TRUE,
  show = FALSE,
  title = TRUE,
  folder = getwd(),
  scenRef = NULL,
  scenDiff = NULL,
  xRef = NULL,
  xDiff = NULL,
  diff_type = "both",
  diff_type_x = "both",
  diff_text_percent = "diffPrcnt",
  diff_text_absolute = "diffAbs",
  diff_text_percent_x = "xdiffPrcnt",
  diff_text_absolute_x = "xdiffAbs",
  width = NULL,
  height = NULL,
  append = "",
  break_interval = NULL,
  include_points = FALSE
)

Arguments

data

Default = NULL. Dataframe to test and convert.

col_agg

Default = "class". Column to remove and then aggregate by.

chart_type

Default = "all". Choices one or more of "all", "param_absolute", "class_absolute", "param_diff_absolute", "class_diff_absolute", "param_diff_percent", "class_diff_percent", "region_absolute"

aspect_ratio

Default = 0.75. aspect ratio

size

Default = 1.5. line size

size_text

Default = 15. Text size

theme

Default = NULL.

ncol

Default = 3. Number of columns.

scales

Default = "free". Choose between "free", "free_y", "free_x", "fixed"

save

Default = TRUE. Save plots.

show

Default = FALSE. Whether to show figure outputs in console.

title

Default = TRUE. Used for region or subRegions or to add title to all plots.

folder

Default = getwd().

scenRef

Default = NULL. Reference Scenario

scenDiff

Default = NULL. Difference Scenarios

xRef

Default = NULL. Reference x

xDiff

Default = NULL. Difference x

diff_type

Default = "both". One of "absolute", "percent", "both".

diff_type_x

Default = "both". One of "absolute", "percent", "both".

diff_text_percent

Default = "_diffPrcnt"

diff_text_absolute

Default = "_diffAbs"

diff_text_percent_x

Default = "_xdiffPrcnt"

diff_text_absolute_x

Default = "_xdiffAbs"

width

Default = NULL

height

Default = NULL

append

Default = ""

break_interval

Default = NULL. Intervals between x breaks starting from first x point.

include_points

Default = FALSE. Add data points to all line charts.